requirements.txt
# Python dependencies
scapy>=2.4.5
scikit-learn>=1.0.2
pyyaml>=6.0
requests>=2.26.0
joblib>=1.1.0